+++ src/Pages/Search/SearchPage.tsx
@@ -2,6 +2,9 @@ import React, { useEffect, useState } from 'react';
 import { requestApi } from '../../Util/api';
 import { useLocation } from 'react-router-dom';
 import { MovieProps } from '../../Util/interface';
+import MovieItem from '../../Components/Movie/MovieItem';
+import { SearchWrap } from './SearchPage.style';
+import { useDebounce } from '../../Hooks/useDebounce';
 
 const SearchPage = () => {
   const [searchResults, setSearchResults] = useState<MovieProps[] | null>(
@@ -9,6 +12,7 @@ const SearchPage = () => {
   );
   const query = new URLSearchParams(useLocation().search);
   const searchTerm = query.get('q');
+  const debouncedSearchTerm = useDebounce(searchTerm || '', 500);
   console.log(searchTerm);
 
   /** 검색어 query string 결과 fetch */
@@ -24,13 +28,22 @@ const SearchPage = () => {
   };
 
   useEffect(() => {
-    requestSearchMovie(searchTerm);
-  }, []);
+    if (debouncedSearchTerm) requestSearchMovie(debouncedSearchTerm);
+  }, [debouncedSearchTerm]);
 
+  console.log(searchResults);
   return (
-    <div>
-      {searchResults && searchResults.map((i) => <p>{i.title}</p>)}
-    </div>
+    <SearchWrap>
+      {searchResults && searchResults.length !== 0 ? (
+        searchResults.map((movie, idx) => (
+          <MovieItem key={idx} movie={movie} />
+        ))
+      ) : (
+        <p className='no-results'>
+          "{searchTerm}" 에 맞는 결과가 없습니다.
+        </p>
+      )}
+    </SearchWrap>
   );
 };
